REFACTORING

+++ model/rbac_options.go
@@ -2,5 +2,5 @@ package model
 
 type RbacOptions struct {
 	Groups []int
-	Access uint32
+	Access uint8
 }
